requirements.txt
google-genai==1.7.0
python-dotenv==1.0.1
gradio==5.22.0